requirements.txt
google-generativeai
pandas
numpy
fastapi
python-dotenv
openpyxl
uvicorn
azure-storage-blob
tqdm